Splitted cooling/color tests to different modules
